requirements.txt
flask==3.0.3
werkzeug==3.0.3
pydantic==2.8.2
python-multipart==0.0.9
pillow==10.4.0
pdfminer.six==20240706
pytesseract==0.3.13
Jinja2==3.1.4
reportlab==4.2.2
openai>=1.37.0
python-docx==1.1.2
